Seperated the each merge of classifications because want there to be a check at the end of each that flags if the merge has meant the number of observations have changed. A task I will work on is integrate a number of the classifcations into one. Additionally have excluded table 14 BEC for now. Below is a step by step process highlighting the remaining challenges.
1) After cleaning raw data. The DataFrame "tradeStatsCommodities" has 12540 observations, and 37 variables

2) After Mode of Transport classifcations merged, 12540 observations, and 38 variables

3) After HS Code classifications merged, 12540 observations, and 39 variables

4) After SITC classifications merged, 12538 observations, 39 variables. Two observations have dropped, investigated this and was because there were 2 null values in raw data-SITC column stored as "NULL" and "null". Have to run a check in the cleaning section that looks for null values maybe, this can be then flagged up to us.

5) After Country Description Imports classifcations merged, 12538 observations and 45 variables

6) After Country Description Exports classifcations merged, 12538 observations and 50 variables

7) After Principle Commidity classifciations merged, 12542 observations and 56 variables. Four observations have be added. Not sure how to run a check against the DataFrames of cleaned data and classifications to understand how we have gained four observations.

Have excluded at this stage merging the BEC classifications because of a number of issues with this information. Below arer the list of tables with any oustanding issues. The only table that has code to add dataframe into an excel spreadhseet is for table number one Balance of Trade.

Table 1: Balance of Trade
Figures produced in R checked against excel and correct. Additionally checked against Annas January calculation and is correct too.

Table 2: Import by HS Code
Figures produced in R checked against excel and correct. Additionally checked against Annas January calculation and is correct too.

Table 3: Export by HS Code
Figures produced in R checked against excel and correct. Additionally checked against Annas January calculation and is correct too.

Table 4: Re-Export by HS Code
Figures produced in R checked against excel and correct. Additionally checked against Annas January calculation and is correct too.

Table 5: Total Export by HS Code
Figures produced in R checked against excel and correct. Additionally checked against Annas January calculation and is correct too.

The rest of the tables haven't been able to check against Annas figures yet but will. I suspect this is where the issues will be though due to the classifications.

Table 6: Principle Exports
Discrepancy between excel total of stats value 296464118, and R of 296448733. A difference here of 15385. After investigation found that a fuel was misclassified and added into exports. Need to investigate further with Customs.

Discrepancy in Re-exports in excel total is 132323734 and in R 132309563, a difference here of 14171. After further investigation the difference is due to "Wood and artciles thereof", it is in excel but not R. I think issue here is that in principle commodities classifications have to add a column specifically for re-exports. Will do this, check again and create a new branch.

Can't add the two dataframes togther "Exports" and "Re-exports" as have a different number of variables in each- gives an error.

Table 7: Principle Imports
The code is not running, am getting error "undefined columns selected". Went through all code and found some issues such as white spaces and full stops in the classifcations. Have changed these and still getting the same error though. Will keep on investigating, am assuming must be something with connecting the list to the dataframe. A descrepency in the lanaguage.

Table 8: Balance of Trade by Major Partner Countries
Large descrepency between total exports. In R was 313983644, and in excel 428787852, a difference of 114804208. After investigation is because of "Other Countries", currently all countries have a code assigned to them. What we have to do is for all countries that are not our principle reporting ones (China, Australia etc) we have to write code to group them together as "Other Countries". Can see this as Other Countries always comes out as 0 stats value (no countries are assigned to this).

This is a similar issue with Balance of Trade by Major Partner countries for "Imports". The other countries has no assigned value. The value in excel is 3096312144, and in R 2928393896.

A further comment would be for future discusions, are there are a number of countries we trade a large amount with that are grouped as Other including Tuvalu, Wallis and Futuna Islands etc

Table 9: Balance of Trade by Region
Checked aginst excel figures and were correct, as stated above have not been able to check against Annas. However have an error when calculating Balalnce of Trade for the Pacific Islands. I think the error is due to using the variables that are within the list- they are not seperate vectors, therefore subtracting them will not work. Am not sure how to resolve this issue though.

Table 10: Trade by Mode of Transport
Small issue with the coding after investigation. Have used "LAND" as a classification in R. However LAND is defined at "OTHER" in the excel classifications. Will recify this and rerun.

Table 11: Trade by Trade Agreement
Not sure how to write code for this table, Anna has stated what she would like. The variables we are interested in are Imports (4000, 4071, 7100), PRF (MSG), SITC (Codes 0-9) and Prnciple Commodities. I will highlight this in an excel and hope that will help in thinking through how to write code for it.

Table 12: Exports by SITC
Checked aginst excel figures and were correct, as stated above have not been able to check against Annas.

Table 13: Imports by SITC
Checked aginst excel figures and were correct, as stated above have not been able to check against Annas.

Table 14: On Hold (BEC)

Summary Issues support from Joe
i) Code to flag if the merge has meant the number of observations have changed

ii) How to run a check in the cleaning section that looks for null values

iii) Run a check against DataFrames of cleaned data and Principle Commodities classifications to understand how we have gained four observations

iv) Table 6- Can't add the two dataframes togther "Exports" and "Re-exports" as have a different number of variables in each- gives an error

v) Table 8- have to write code to group countries which are not our principle ones as "Other". Currently other countries has no assigned value and so always equal to 0

vi) Table 9- calculating the balalnce of Trade for the Pacific Islands, think error is due to using the variables that are within the list- they are not seperate vectors, therefore subtracting them will not work.

vii) Table 11-Not sure how to write code for this table, will send a template in Excel